ingress: add nginx ingress controller images configuration API
nginx ingress controller image and webhook certgen image are hard-coded in the previous version,
which is not friendly to users, so add the API for users to config the related images locations
themselves, which means users can upgrade the nginx ingress controller as well.

Note: for the nginx ingress controller version has a relationship with the kubernetes version,
when users config the nginx ingress controller image, they should ensure these 2 versions are
matchable.

Signed-off-by: zzguang <zhengguang.zhang@intel.com>