ingress: add YurtIngress enhancement features (#51)
* ingress: remove singleton restriction for yurtingress in webhook

Signed-off-by: Linda Yu <linda.yu@intel.com>

* ingress: remove singleton yurtingress CR restriction

Signed-off-by: zhenggu1 <zhengguang.zhang@intel.com>

* ingress: add ingress service externalIPs API

When the ingress controller service for a nodepool is exposed, user can set
the service external IP by the new added API in YurtIngress CR.

Signed-off-by: zzguang <zhengguang.zhang@intel.com>

* ingress: add nginx ingress controller images configuration API

nginx ingress controller image and webhook certgen image are hard-coded in the previous version,
which is not friendly to users, so add the API for users to config the related images locations
themselves, which means users can upgrade the nginx ingress controller as well.

Note: for the nginx ingress controller version has a relationship with the kubernetes version,
when users config the nginx ingress controller image, they should ensure these 2 versions are
matchable.

Signed-off-by: zzguang <zhengguang.zhang@intel.com>

* ingress: update nginx ingress controller go template

Signed-off-by: zzguang <zhengguang.zhang@intel.com>

* ingress: fix ingress controller resources cleanup issue

When yurt-app-manager is deleted, some nginx ingress controller related
resources can not be cleaned up or garbage collected, add ownerReferences
to these resources to avoid this issue.

Signed-off-by: zzguang <zhengguang.zhang@intel.com>

* ingress: move YurtIngress CR default values to mutating webhook

Signed-off-by: zzguang <zhengguang.zhang@intel.com>

Co-authored-by: Linda Yu <linda.yu@intel.com>